Seperate ui generation from information display
Implemented mechanism via which the full user interface of the
application is generated upon starting the project. Additionally made an
implemention of displaying the local time information of a location when
searched for and create a mechanism for a clock running for the actual
local time of the searched location.

Created mechanism via which to auto generate a new background depending
on the time and weather conditions of a location that was search for.
Ensured to seperate the ui generation functionality from the information
displaying mechanisms.